Still broke

+++ AoSSimCPP/Weapon.cpp
@@ -2,12 +2,12 @@
 #include "Weapon.h"
 
 Weapon::Weapon(std::string name, 
-	uint8_t range, 
-	uint8_t attacks, 
-	uint8_t tohit, 
-	uint8_t towound, 
-	uint8_t rend, 
-	uint8_t damage) : GameEntity(name, ""), myStats(range, attacks, tohit, towound, rend, damage)
+	size_t range, 
+	size_t attacks, 
+	size_t tohit, 
+	size_t towound, 
+	size_t rend, 
+	size_t damage) : GameEntity(name, ""), myStats(range, attacks, tohit, towound, rend, damage)
 {
 }
 
@@ -32,13 +32,13 @@ void Weapon::EndTurn()
 
 }
 
-int Weapon::MakeRoll(uint8_t target)
+int Weapon::MakeRoll(size_t target)
 {
 	return (Die::Roll() >= target);
 }
 
 //This is longer than it needs to be right now, as I need it to be clean for possible expansion when I add abilities etc.
-uint8_t Weapon::GenerateWounds(uint8_t save)
+size_t Weapon::GenerateWounds(size_t save)
 {
 	int hits, wounds, saves;
 	hits = wounds = saves = 0;
@@ -46,5 +46,5 @@ uint8_t Weapon::GenerateWounds(uint8_t save)
 	for (int i = 0; i < hits; i++) wounds += MakeRoll(myStats.towound);
 	for (int i = 0; i < wounds; i++) saves += ((Die::Roll() + myStats.rend) >= save);
 
-	return (uint8_t) max(0, (wounds - saves) * myStats.damage);
+	return (size_t) max(0, (wounds - saves) * myStats.damage);
 }
